ready for commentary
