+++ docs/docs/services/apigatewayv2.rst
@@ -33,10 +33,10 @@ apigatewayv2
         CredentialsArn, RouteKey, Tags, Target
         
 
-- [ ] create_api_mapping
+- [X] create_api_mapping
 - [X] create_authorizer
 - [ ] create_deployment
-- [ ] create_domain_name
+- [X] create_domain_name
 - [X] create_integration
 - [X] create_integration_response
 - [X] create_model
@@ -66,8 +66,8 @@ apigatewayv2
 - [X] delete_vpc_link
 - [ ] export_api
 - [X] get_api
-- [ ] get_api_mapping
-- [ ] get_api_mappings
+- [X] get_api_mapping
+- [X] get_api_mappings
 - [X] get_apis
   
         Pagination is not yet implemented
@@ -77,8 +77,12 @@ apigatewayv2
 - [ ] get_authorizers
 - [ ] get_deployment
 - [ ] get_deployments
-- [ ] get_domain_name
-- [ ] get_domain_names
+- [X] get_domain_name
+- [X] get_domain_names
+  
+        Pagination is not yet implemented
+        
+
 - [X] get_integration
 - [X] get_integration_response
 - [X] get_integration_responses
