+++ docs/docs/services/ecs.rst
@@ -60,6 +60,7 @@ ecs
 - [X] describe_tasks
 - [ ] discover_poll_endpoint
 - [ ] execute_command
+- [ ] get_task_protection
 - [X] list_account_settings
 - [X] list_attributes
   
@@ -110,5 +111,6 @@ ecs
 - [X] update_service_primary_task_set
   Updates task sets be PRIMARY or ACTIVE for given cluster:service task sets
 
+- [ ] update_task_protection
 - [X] update_task_set
