+++ docs/docs/services/emr-containers.rst
@@ -28,14 +28,18 @@ emr-containers
 |start-h3| Implemented features for this service |end-h3|
 
 - [X] cancel_job_run
+- [ ] create_job_template
 - [ ] create_managed_endpoint
 - [X] create_virtual_cluster
+- [ ] delete_job_template
 - [ ] delete_managed_endpoint
 - [X] delete_virtual_cluster
 - [X] describe_job_run
+- [ ] describe_job_template
 - [ ] describe_managed_endpoint
 - [X] describe_virtual_cluster
 - [X] list_job_runs
+- [ ] list_job_templates
 - [ ] list_managed_endpoints
 - [ ] list_tags_for_resource
 - [X] list_virtual_clusters
