+++ docs/docs/services/es.rst
@@ -30,15 +30,18 @@ es
 - [ ] accept_inbound_cross_cluster_search_connection
 - [ ] add_tags
 - [ ] associate_package
+- [ ] authorize_vpc_endpoint_access
 - [ ] cancel_elasticsearch_service_software_update
 - [X] create_elasticsearch_domain
 - [ ] create_outbound_cross_cluster_search_connection
 - [ ] create_package
+- [ ] create_vpc_endpoint
 - [X] delete_elasticsearch_domain
 - [ ] delete_elasticsearch_service_role
 - [ ] delete_inbound_cross_cluster_search_connection
 - [ ] delete_outbound_cross_cluster_search_connection
 - [ ] delete_package
+- [ ] delete_vpc_endpoint
 - [ ] describe_domain_auto_tunes
 - [ ] describe_domain_change_progress
 - [X] describe_elasticsearch_domain
@@ -50,6 +53,7 @@ es
 - [ ] describe_packages
 - [ ] describe_reserved_elasticsearch_instance_offerings
 - [ ] describe_reserved_elasticsearch_instances
+- [ ] describe_vpc_endpoints
 - [ ] dissociate_package
 - [ ] get_compatible_elasticsearch_versions
 - [ ] get_package_version_history
@@ -66,11 +70,16 @@ es
 - [ ] list_elasticsearch_versions
 - [ ] list_packages_for_domain
 - [ ] list_tags
+- [ ] list_vpc_endpoint_access
+- [ ] list_vpc_endpoints
+- [ ] list_vpc_endpoints_for_domain
 - [ ] purchase_reserved_elasticsearch_instance_offering
 - [ ] reject_inbound_cross_cluster_search_connection
 - [ ] remove_tags
+- [ ] revoke_vpc_endpoint_access
 - [ ] start_elasticsearch_service_software_update
 - [ ] update_elasticsearch_domain_config
 - [ ] update_package
+- [ ] update_vpc_endpoint
 - [ ] upgrade_elasticsearch_domain
